Use %s string formatting in system/timezone.py
This module was using python 2.6 string .format().
To enable the module to run on python2.4, replace
the .format formatting with '%s' based string formatting.

There was also a use of a 'filename' variable in the
NosystemdTimezone.get() method that was never set.

An import of 'os' was also added for clarity.